backend/verification_service/requirements.txt
grpcio==1.66.1
grpcio-tools==1.66.1
httpx==0.27.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
python-dotenv==1.0.1
